bagurush/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-huggingface>=0.1.0
faiss-cpu>=1.7.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
pydantic>=2.0.0
pypdf>=3.0.0
unstructured>=0.10.0
sentence-transformers>=2.0.0
streamlit>=1.28.0
pytest>=7.0.0

